# src/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "typeorm-rest-example",
		"description": "This is a example REST API using TypeORM.",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		},
		"version": "1.0.0"
	},
...
	"servers": [
		{
			"url": "http://localhost:4000"
		}
	],
...
	"paths": {
		"/api/users": {
			"get": {
				"tags": ["Users", "Admin"],
				"summary": "Admin Get all users",
				"description": "This can only be done by an isAdmin user.",
				"operationId": "getAllUsers",
				"parameters": [
					{
						"name": "search",
						"in": "query",
						"description": "Search from firstName, lastName and email",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Limit number of items in the query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"description": "Offset number of items in the query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "sortDirection",
						"in": "query",
						"description": "Sorting order of the Id",
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "All users in the system",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/UserFromDB"
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				},
				"x-codegen-request-body-name": "body"
			},
			"post": {
				"tags": ["Users"],
				"summary": "Register a user",
				"description": "Register a user on the platform.",
				"operationId": "registerUser",
				"requestBody": {
					"description": "User object required for registering",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "successfully registered user",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer"
										},
										"firstName": {
											"type": "string",
											"format": "string"
										},
										"lastName": {
											"type": "string",
											"format": "string"
										},
										"email": {
											"type": "string",
											"format": "email"
										},
										"isAdmin": {
											"type": "boolean",
											"default": false
										},
										"isEmailConfirmed": {
											"type": "boolean",
											"default": false
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				},
				"x-codegen-request-body-name": "body"
			}
		},
...
		"/api/users/refreshToken": {
			"get": {
				"tags": ["Users"],
				"summary": "Get new access-token via refresh-token in cookies",
				"operationId": "refreshAuth",
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StandardError"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/users/confirmUser": {
			"post": {
				"tags": ["Users"],
				"summary": "Request new account confirmation email",
				"operationId": "resendConfirmationEmail",
				"requestBody": {
					"description": "Pass email to check and send email from db. Will always return true for security purposes",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Users"],
				"summary": "Confirms the user email into the system",
				"operationId": "confirmUser",
				"requestBody": {
					"description": "Confirm user email",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"token": {
										"type": "string"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserForApplication"
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					},
					"401": {
						"description": "Invalid token or user not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StandardError"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/changeEmail": {
			"post": {
				"tags": ["Users"],
				"summary": "Request confirmation email be sent to verify new email",
				"operationId": "sendChangeEmailConfirmation",
				"security": [{ "bearerAuth": [] }],
				"requestBody": {
					"description": "Send the email address to replace the existing one so a confirmation can be sent.",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					},
					"401": {
						"description": "Invalid token or user not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StandardError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Users"],
				"summary": "Confirms the new user email into the system",
				"operationId": "confirmNewUserEmail",
				"requestBody": {
					"description": "Confirms the new user email",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"token": {
										"type": "string"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserForApplication"
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/resetPassword": {
			"post": {
				"tags": ["Users"],
				"summary": "Send reset password link to email",
				"operationId": "sendForgotPasswordEmail",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Users"],
				"summary": "Change user password using the token from the email",
				"operationId": "resetPasswordWithToken",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": true,
								"properties": {
									"token": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"FieldValidationError": {
				"type": "object",
				"properties": {
					"errors": {
						"type": "object",
						"properties": {
							"body": {
								"type": "array",
								"nullable": true,
								"format": "possibleVoid",
								"items": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										},
										"propertyPath": {
											"type": "string"
										}
									}
								}
							},
							"query": {
								"type": "array",
								"nullable": true,
								"format": "possibleVoid",
								"items": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										},
										"propertyPath": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"StandardError": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string",
						"nullable": true
					}
				}
			},
			"SortDirection": {
				"type": "string",
				"enum": ["ASC", "DESC"]
			},
			"UserFromDB": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"password": {
						"type": "string"
					},
					"isEmailConfirmed": {
						"type": "boolean"
					},
					"isAdmin": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					}
				}
			},
			"UserForApplication": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"isEmailConfirmed": {
						"type": "boolean"
					},
					"isAdmin": {
						"type": "boolean"
					}
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}